requirements.txt
asgiref==3.8.1
awscli-local==0.22.0
boto3==1.37.36
botocore==1.37.36
Django==5.2
djangorestframework==3.16.0
jmespath==1.0.1
lark==1.2.2
localstack-client==2.10
packaging==24.2
psycopg2-binary==2.9.10
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-hcl2==7.1.0
s3transfer==0.11.5
six==1.17.0
sqlparse==0.5.3
terraform-local==0.21.0
typing_extensions==4.13.2
urllib3==2.4.0
